api refactor in one endpoint with more parameters
